# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShaadiHub Wedding Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  vendor: "vendor",
  admin: "admin",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export type SubscriptionPlan =
  (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan];

export const SubscriptionPlan = {
  free: "free",
  premium: "premium",
} as const;

export type VendorCategory =
  (typeof VendorCategory)[keyof typeof VendorCategory];

export const VendorCategory = {
  Marriage_Hall: "Marriage Hall",
  Catering: "Catering",
  Photography: "Photography",
  Videography: "Videography",
  Decoration: "Decoration",
  Car_Rental: "Car Rental",
  Event_Planner: "Event Planner",
  "Makeup_&_Beauty": "Makeup & Beauty",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  isBlocked: boolean;
  createdAt: string;
}

export interface VendorImage {
  id: number;
  vendorId: number;
  imageUrl: string;
  publicId: string;
}

export interface Package {
  id: number;
  vendorId: number;
  name: string;
  description?: string | null;
  price: number;
  features: string[];
  isPopular: boolean;
}

export interface Vendor {
  id: number;
  userId: number;
  category: VendorCategory;
  title: string;
  description?: string | null;
  location?: string | null;
  area?: string | null;
  city: string;
  latitude?: number | null;
  longitude?: number | null;
  priceRangeMin?: number | null;
  priceRangeMax?: number | null;
  capacity?: number | null;
  whatsapp?: string | null;
  coverImageUrl?: string | null;
  verified: boolean;
  isFeatured: boolean;
  isSuspended: boolean;
  rating: number;
  totalReviews: number;
  profileViews: number;
  createdAt: string;
  images: VendorImage[];
  packages: Package[];
}

export interface Review {
  id: number;
  vendorId: number;
  userId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: string;
  userName?: string | null;
  userAvatarUrl?: string | null;
}

export interface Booking {
  id: number;
  vendorId: number;
  userId: number;
  packageId?: number | null;
  eventDate: string;
  guestCount?: number | null;
  notes?: string | null;
  status: BookingStatus;
  createdAt: string;
  vendorTitle?: string | null;
  vendorCategory?: string | null;
  vendorCoverImageUrl?: string | null;
  packageName?: string | null;
  userName?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  totalVendors: number;
  verifiedVendors: number;
  pendingVendors: number;
  totalBookings: number;
  totalReviews: number;
}

export interface VendorListResponse {
  total: number;
  page: number;
  pages: number;
  vendors: Vendor[];
}

export interface AuthResponse {
  token: string;
  user: User;
  vendorId?: number | null;
}

export interface FavoriteToggleResponse {
  favorited: boolean;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBody = {
  name: string;
  email: string;
  password: string;
  role: UserRole;
  phone?: string;
  category?: VendorCategory;
  title?: string;
};

export type LoginBody = {
  email: string;
  password: string;
};

export type ListVendorsParams = {
  category?: string;
  area?: string;
  city?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  capacity?: number;
  search?: string;
  featured?: boolean;
  page?: number;
  limit?: number;
};

export type CreateVendorBody = {
  category: VendorCategory;
  title: string;
  description?: string;
  location?: string;
  area?: string;
  city?: string;
  latitude?: number;
  longitude?: number;
  priceRangeMin?: number;
  priceRangeMax?: number;
  capacity?: number;
  whatsapp?: string;
};

export type UpdateVendorBody = {
  category?: VendorCategory;
  title?: string;
  description?: string;
  location?: string;
  area?: string;
  city?: string;
  latitude?: number;
  longitude?: number;
  priceRangeMin?: number;
  priceRangeMax?: number;
  capacity?: number;
  whatsapp?: string;
};

export type UploadVendorImageBody = {
  file: Blob;
};

export type CreatePackageBody = {
  name: string;
  description?: string;
  price: number;
  features?: string[];
  isPopular?: boolean;
};

export type UpdatePackageBody = {
  name?: string;
  description?: string;
  price?: number;
  features?: string[];
  isPopular?: boolean;
};

export type UpdateMeBody = {
  name?: string;
  phone?: string;
  avatarUrl?: string;
};

export type CreateBookingBody = {
  vendorId: number;
  packageId?: number;
  eventDate: string;
  guestCount?: number;
  notes?: string;
};

export type CreateReviewBody = {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
};

export type UpdateBookingStatusParams = {
  status: BookingStatus;
};