# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShaadiHub Wedding Marketplace API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRole:
      type: string
      enum: [user, vendor, admin]
    BookingStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
...
    VendorCategory:
      type: string
      enum:
        - Marriage Hall
        - Catering
        - Photography
        - Videography
        - Decoration
        - Car Rental
        - Event Planner
        - Makeup & Beauty
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - isBlocked
        - createdAt
    VendorImage:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        imageUrl:
          type: string
        publicId:
          type: string
      required:
        - id
        - vendorId
        - imageUrl
        - publicId
    Package:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        features:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
      required:
        - id
        - vendorId
        - name
        - price
        - features
        - isPopular
    Vendor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          $ref: "#/components/schemas/VendorCategory"
        title:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        city:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        priceRangeMin:
          type: number
          nullable: true
        priceRangeMax:
          type: number
          nullable: true
        capacity:
          type: integer
          nullable: true
        whatsapp:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        verified:
          type: boolean
        isFeatured:
          type: boolean
        isSuspended:
          type: boolean
        rating:
          type: number
        totalReviews:
          type: integer
        profileViews:
          type: integer
        createdAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/VendorImage"
        packages:
          type: array
          items:
            $ref: "#/components/schemas/Package"
      required:
        - id
        - userId
        - category
        - title
        - city
        - verified
        - isFeatured
        - isSuspended
        - rating
        - totalReviews
        - profileViews
        - createdAt
        - images
        - packages
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        userId:
          type: integer
        packageId:
          type: integer
          nullable: true
        eventDate:
          type: string
          format: date
        guestCount:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookingStatus"
        createdAt:
          type: string
          format: date-time
        vendorTitle:
          type: string
          nullable: true
        vendorCategory:
          type: string
          nullable: true
        vendorCoverImageUrl:
          type: string
          nullable: true
        packageName:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
      required:
        - id
        - vendorId
        - userId
        - eventDate
        - status
        - createdAt
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalVendors:
          type: integer
        verifiedVendors:
          type: integer
        pendingVendors:
          type: integer
        totalBookings:
          type: integer
        totalReviews:
          type: integer
      required:
        - totalUsers
        - totalVendors
        - verifiedVendors
        - pendingVendors
        - totalBookings
        - totalReviews
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        vendorId:
          type: integer
          nullable: true
      required:
        - token
        - user
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register user or vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  $ref: "#/components/schemas/UserRole"
                phone:
                  type: string
                category:
                  $ref: "#/components/schemas/VendorCategory"
                title:
                  type: string
              required:
                - name
                - email
                - password
                - role
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # VENDORS
...
  /bookings/vendor:
    get:
      operationId: getVendorBookings
      tags: [bookings]
      summary: Get all bookings for logged-in vendor
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vendor bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{id}/status:
    put:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ADMIN
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/vendors:
    get:
      operationId: adminListVendors
      tags: [admin]
      summary: List all vendors (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/approve:
    put:
      operationId: approveVendor
      tags: [admin]
      summary: Approve vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/suspend:
    put:
      operationId: suspendVendor
      tags: [admin]
      summary: Toggle vendor suspension
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspension toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}/feature:
    put:
      operationId: featureVendor
      tags: [admin]
      summary: Toggle vendor featured status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Featured toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /admin/vendors/{id}:
    delete:
      operationId: adminDeleteVendor
      tags: [admin]
      summary: Delete vendor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/block:
    put:
      operationId: blockUser
      tags: [admin]
      summary: Toggle user block status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Block toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"