# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShaadiHub Wedding Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  vendorId: number;
  userId: number;
  packageId?: number | null;
  eventDate: Date;
  guestCount?: number | null;
  notes?: string | null;
  status: BookingStatus;
  createdAt: Date;
  vendorTitle?: string | null;
  vendorCategory?: string | null;
  vendorCoverImageUrl?: string | null;
  packageName?: string | null;
  userName?: string | null;
}