# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShaadiHub Wedding Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;