# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ShaadiHub Wedding Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Package } from "./package";
import type { VendorCategory } from "./vendorCategory";
import type { VendorImage } from "./vendorImage";

export interface Vendor {
  id: number;
  userId: number;
  category: VendorCategory;
  title: string;
  description?: string | null;
  location?: string | null;
  area?: string | null;
  city: string;
  latitude?: number | null;
  longitude?: number | null;
  priceRangeMin?: number | null;
  priceRangeMax?: number | null;
  capacity?: number | null;
  whatsapp?: string | null;
  coverImageUrl?: string | null;
  verified: boolean;
  isFeatured: boolean;
  isSuspended: boolean;
  rating: number;
  totalReviews: number;
  profileViews: number;
  createdAt: Date;
  images: VendorImage[];
  packages: Package[];
}